js_fix_cleaning: refactored view architecture
