--- src/app/model/form-elements/InputTextElement.ts
@@ -1,6 +1,0 @@
-export interface InputTextElement {
-  fieldTitle: string,
-  formControlName: string;
-  readOnly?: boolean;
-  initialValue?: string;
-}
