+++ src/app/views/admin/admin-views/students/student-details/student-details.component.html
