+++ src/app/views/default/landing/landing.component.html
@@ -1,7 +1,7 @@
 
 <div class="grid grid-nogutter surface-section text-800">
   <div class="col-12 md:col-6 overflow-hidden">
-    <img src="assets/img/landing/car.jpg" alt="Car" class="md:ml-auto block md:h-full">
+    <img src="../../../../assets/img/landing/car.jpg" alt="Car" class="md:ml-auto block md:h-full">
   </div>
   <div class="col-12 md:col-6 p-6 text-center md:text-left flex align-items-center ">
     <section>
