+++ src/app/views/default/landing/landing.component.ts
